购物车、支付、订单生成、地址管理

+++ eshop/src/main/java/com/shop/eshop/mapper/GoodMapper.java
@@ -13,7 +13,7 @@ public interface GoodMapper {
 
     List<Good> getAllGood();
 
-    Good getGoodById(Integer id);
+    Good getGoodById(Long id);
 
     List<Good> getGoodBySaleNum();
