+++ eshop/src/main/java/com/shop/eshop/service/GoodService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface GoodService {
     List<Good> getAllGood();
-    Good getGoodById(Integer id);
+    Good getGoodById(Long id);
     List<Good> getGoodBySaleNum();
 
     List<Good> getGoodByCondition(TypeCondition tp);
